djangocms_news/requirements.txt
django-multiupload
